+++ oximeter/db/src/model.rs
@@ -1880,7 +1880,6 @@ mod tests {
     use super::*;
     use chrono::Timelike;
     use oximeter::histogram::Record;
-    use oximeter::test_util;
     use oximeter::Datum;
 
     #[test]
@@ -1983,7 +1982,7 @@ mod tests {
 
     #[test]
     fn test_unroll_from_source() {
-        let sample = test_util::make_sample();
+        let sample = oximeter_test_utils::make_sample();
         let out = unroll_from_source(&sample);
         assert_eq!(out["oximeter.fields_string"].len(), 2);
         assert_eq!(out["oximeter.fields_i64"].len(), 1);
@@ -2003,8 +2002,8 @@ mod tests {
     // datum.
     #[test]
     fn test_unroll_missing_measurement_row() {
-        let sample = test_util::make_sample();
-        let missing_sample = test_util::make_missing_sample();
+        let sample = oximeter_test_utils::make_sample();
+        let missing_sample = oximeter_test_utils::make_missing_sample();
         let (table_name, row) = unroll_measurement_row(&sample);
         let (missing_table_name, missing_row) =
             unroll_measurement_row(&missing_sample);
@@ -2022,7 +2021,7 @@ mod tests {
 
     #[test]
     fn test_unroll_measurement_row() {
-        let sample = test_util::make_hist_sample();
+        let sample = oximeter_test_utils::make_hist_sample();
         let (table_name, row) = unroll_measurement_row(&sample);
         assert_eq!(table_name, "oximeter.measurements_histogramf64");
         let unpacked: HistogramF64MeasurementRow =
