+++ oximeter/oximeter/Cargo.toml
@@ -13,9 +13,10 @@ anyhow.workspace = true
 clap.workspace = true
 chrono.workspace = true
 omicron-workspace-hack.workspace = true
-oximeter-impl.workspace = true
 oximeter-macro-impl.workspace = true
+oximeter-schema.workspace = true
 oximeter-timeseries-macro.workspace = true
+oximeter-types.workspace = true
 prettyplease.workspace = true
 syn.workspace = true
 toml.workspace = true
