+++ oximeter/schema/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "oximeter-schema"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+chrono.workspace = true
+heck.workspace = true
+omicron-workspace-hack.workspace = true
+oximeter-types.workspace = true
+proc-macro2.workspace = true
+quote.workspace = true
+schemars.workspace = true
+serde.workspace = true
+slog-error-chain.workspace = true
+syn.workspace = true
+toml.workspace = true
