+++ oximeter/schema/src/lib.rs
@@ -1,0 +1,12 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// Copyright 2024 Oxide Computer Company
+
+//! Tools for working with schemas for fields and timeseries.
+//!
+//! The actual schema type definitions are in [`oximeter_types::schema`].
+
+pub mod codegen;
+pub mod ir;
