+++ oximeter/test-utils/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "oximeter-test-utils"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+chrono.workspace = true
+omicron-workspace-hack.workspace = true
+oximeter-macro-impl.workspace = true
+oximeter-types.workspace = true
+uuid.workspace = true
