+++ oximeter/timeseries-macro/Cargo.toml
@@ -8,7 +8,8 @@ proc-macro = true
 
 [dependencies]
 omicron-workspace-hack.workspace = true
-oximeter-impl.workspace = true
+oximeter-schema.workspace = true
+oximeter-types.workspace = true
 proc-macro2.workspace = true
 quote.workspace = true
 syn.workspace = true
