+++ oximeter/types/src/lib.rs
@@ -4,8 +4,6 @@
 
 // Copyright 2024 Oxide Computer Company
 
-pub use oximeter_macro_impl::*;
-
 // Export the current crate as `oximeter`. The macros defined in `oximeter-macro-impl` generate
 // code referring to symbols like `oximeter::traits::Target`. In consumers of this crate, that's
 // fine, but internally there _is_ no crate named `oximeter`, it's just `self` or `crate`.
@@ -17,15 +15,18 @@ extern crate self as oximeter;
 pub mod histogram;
 pub mod quantile;
 pub mod schema;
-pub mod test_util;
 pub mod traits;
 pub mod types;
 
 pub use quantile::Quantile;
 pub use quantile::QuantileError;
+pub use schema::AuthzScope;
 pub use schema::FieldSchema;
+pub use schema::FieldSource;
+pub use schema::TimeseriesDescription;
 pub use schema::TimeseriesName;
 pub use schema::TimeseriesSchema;
+pub use schema::Units;
 pub use traits::Metric;
 pub use traits::Producer;
 pub use traits::Target;
